+++ pytest_bdd/types.py
@@ -8,6 +8,7 @@ EXAMPLES_HEADERS = 'example headers'  # pragma: no cover
 EXAMPLE_LINE = 'example line'  # pragma: no cover
 EXAMPLE_LINE_VERTICAL = 'example line vertical'  # pragma: no cover
 SCENARIO = 'scenario'  # pragma: no cover
+BACKGROUND = 'background'  # pragma: no cover
 GIVEN = 'given'  # pragma: no cover
 WHEN = 'when'  # pragma: no cover
 THEN = 'then'  # pragma: no cover